added google authentication and user profile

+++ src/main/java/edu/vhhs/demo/vendingmachine/controller/ProductController.java
@@ -3,8 +3,16 @@ package edu.vhhs.demo.vendingmachine.controller;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +20,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import edu.vhhs.demo.vendingmachine.entity.Product;
 import edu.vhhs.demo.vendingmachine.projection.ProductJoin;
@@ -21,11 +30,15 @@ import edu.vhhs.demo.vendingmachine.service.ProductService;
 @RequestMapping(path = "/api/v${ApiVersion}/product")
 public class ProductController {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(ProductController.class);
     @Autowired
-    ProductService productService;
+    @Qualifier("productService")
+    private ProductService productService;
 
     @GetMapping(value = "", produces = MediaType.APPLICATION_JSON_VALUE)
     private List<Product> getAllProduct() {
+        // Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        // LOGGER.info(auth.getAuthorities().toString());
         return productService.getAllProduct();
     }
 
@@ -37,14 +50,22 @@ public class ProductController {
     }
 
     @PostMapping(value = "", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    private Product saveProduct(@RequestBody Product product) {
-        productService.saveOrUpdate(product);
-        return product;
+    private ResponseEntity<Product> saveProduct(@RequestBody Product product) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN_ROLE")))
+            productService.saveOrUpdate(product);
+        else
+            return new ResponseEntity<Product>(product, HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity<Product>(product, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
     private void deleteProduct(@PathVariable("id") int id) {
-        productService.delete(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN_ROLE")))
+            productService.delete(id);
+        else
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
     }
 
     /**
