+++ src/main/java/edu/vhhs/demo/vendingmachine/controller/VendingMachineController.java
@@ -4,6 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,6 +17,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import edu.vhhs.demo.vendingmachine.entity.VendingMachine;
 import edu.vhhs.demo.vendingmachine.service.VendingMachineService;
@@ -20,7 +27,8 @@ import edu.vhhs.demo.vendingmachine.service.VendingMachineService;
 public class VendingMachineController {
 
     @Autowired
-    VendingMachineService vendingMachineService;
+    @Qualifier("vendingMachineService")
+    private VendingMachineService vendingMachineService;
 
     @GetMapping("")
     private List<VendingMachine> getAllVendingMachine() {
@@ -35,13 +43,21 @@ public class VendingMachineController {
     }
 
     @PostMapping("")
-    private VendingMachine saveVendingMachine(@RequestBody VendingMachine vendingMachine) {
-        vendingMachineService.saveOrUpdate(vendingMachine);
-        return vendingMachine;
+    private ResponseEntity<VendingMachine> saveVendingMachine(@RequestBody VendingMachine vendingMachine) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN_ROLE")))
+            vendingMachineService.saveOrUpdate(vendingMachine);
+        else
+            return new ResponseEntity<VendingMachine>(vendingMachine, HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity<VendingMachine>(vendingMachine, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
     private void deleteVendingMachine(@PathVariable("id") int id) {
-        vendingMachineService.delete(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN_ROLE")))
+            vendingMachineService.delete(id);
+        else
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
     }
 }
